requirements.txt
# ============================================================================
# Flask Core & Web Server
# ============================================================================
Flask==2.2.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Babel==3.0.1
gunicorn==21.2.0              # WSGI HTTP Server pour production (Render)
python-dotenv==1.0.0

# ============================================================================
# Database & Storage
# ============================================================================
tinydb==4.7.0                 # NoSQL database (JSON-based, file storage)
SQLAlchemy==1.4.48            # ORM

# ============================================================================
# Scheduling & Task Management
# ============================================================================
APScheduler==3.10.1           # Planification de tâches serveur

# ============================================================================
# Data Processing & Analysis
# ============================================================================
pandas==2.1.2                 # Data manipulation
nltk==3.8.1                   # Natural Language Processing
requests==2.31.0              # HTTP requests

# ============================================================================
# Document Processing
# ============================================================================
python-docx==0.8.11           # Word documents
pdfkit==1.0.0                 # PDF generation

# ============================================================================
# AI/ML (optionnel - peut être désactivé en production)
# ============================================================================
transformers==4.35.0          # Hugging Face transformers
torch>=2.0.0                  # PyTorch

# ============================================================================
# Payment Integration
# ============================================================================
paypalrestsdk==1.14.0         # PayPal SDK

# ============================================================================
# SMS & Communication
# ============================================================================
twilio==8.4.1                 # SMS & WhatsApp

# ============================================================================
# Production Utilities
# ============================================================================
Werkzeug==2.2.3               # WSGI toolkit
Click==8.1.7                  # CLI utilities
MarkupSafe==2.1.3             # String escaping